BACKEND/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-environ==0.12.0
psycopg2-binary==2.9.10
gunicorn==20.1.0
whitenoise==6.6.0